pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "zenopay-sdk"
dynamic = ["version"]
description = "A modern Python SDK for the ZenoPay payment API"
readme = "README.md"
license = "MIT"
authors = [{ name = "Elution Hub", email = "elusion.lab@gmail.com" }]
maintainers = [{ name = "Elution Hub", email = "elusion.lab@gmail.com" }]
keywords = [
    "zenopay",
    "payments",
    "api",
    "ussd",
    "mobile-money",
    "sdk",
    "africa",
    "tanzania",
    "payments-gateway"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial :: Point-Of-Sale",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
requires-python = ">=3.8"
dependencies = [
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "typing-extensions>=4.0.0; python_version < '3.11'",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "respx>=0.20.0",
    "mypy>=1.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "pre-commit>=3.0.0",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.0.0",
    "mkdocstrings[python]>=0.20.0",
]
server = ["flask>=2.0.0", "fastapi>=0.68.0", "uvicorn>=0.15.0"]

[project.urls]
Homepage = "https://github.com/elusionhub/zenopay-python-sdk"
Documentation = "https://github.com/elusionhub/zenopay-python-sdk#readme"
Repository = "https://github.com/elusionhub/zenopay-python-sdk.git"
Issues = "https://github.com/elusionhub/zenopay-python-sdk/issues"
Changelog = "https://github.com/elusionhub/zenopay-python-sdk/blob/main/CHANGELOG.md"

[tool.hatch.version]
path = "src/elusion/zenopay/__init__.py"

[tool.hatch.build.targets.wheel]
packages = ["src/elusion"]

[tool.hatch.build]
include = ["src/elusion"]

[tool.setuptools.package-data]
"zenopay" = ["py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config"
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
]

[tool.coverage.run]
source = ["src/elusion/zenopay"]
omit = ["*/tests/*", "*/examples/*", "*/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.mypy]
python_version = "3.8"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
implicit_reexport = false
strict_equality = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false

[tool.black]
line-length = 150
target-version = ["py38", "py39", "py310", "py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | __pycache__
)/
'''

[tool.flake8]
max-line-length = 150
extend-ignore = ["W391", "E203", "W503"]
exclude = [
    "__pycache__",
    ".git",
    ".pytest_cache",
    "build",
    "dist",
    "*.egg-info",
    ".venv",
    ".mypy_cache"
]

[tool.bumpversion]
current_version = "0.0.1"
commit = true
tag = true
tag_name = "{new_version}"
message = "Bump version: {current_version} → {new_version}"

[[tool.bumpversion.files]]
filename = "src/elusion/zenopay/__init__.py"
